getting the token

+++ ConsoleApplication1/controllers/CompanyUsers.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Formatting;
@@ -21,12 +22,19 @@ namespace OwinSelfhostSample
             var main = new MainAccess();
             var companyUsers = request.Content.ReadAsAsync<List<User>>().Result;
             var companyId = id;
-
+            var idString = request.Headers.GetValues("token").First();
+            var response = new HttpResponseMessage();
+            if (!main.IsUserManager(id, int.Parse(idString)) && !main.CheckIsAdmin(int.Parse(idString)))
+            {
+                response.StatusCode = HttpStatusCode.BadRequest;
+                response.Content = new StringContent("user is not in company");
+                return response;
+            }
             main.DeleteUsersOfCompany(companyId);
 
             main.AddUsersToCompany(companyId,companyUsers);
 
-            var response = new HttpResponseMessage();
+            
             response.StatusCode = HttpStatusCode.OK;
 
             return response;
